[fix] Change package name
